+++ RELEASE NOTES.md
@@ -1,3 +1,11 @@
+Version 1.1
+
+- Added mergeValuesFromObject method.
+- Renamed documentsPath to savePath.
+- Updated loading and saving methods to use the application support folder by default, instead of the documents folder.
+- Fixed nil object exception in loading code.
+- Fixed bug in NSCoded loading logic.
+
 Version 1.0
 
 - Initial release